requirements.txt
jieba~=0.42.1
utils~=1.0.2
snownlp~=0.12.3
pandas~=2.2.3
numpy~=2.2.3
scikit-learn~=1.6.1
PyMySQL~=1.1.1
matplotlib~=3.10.1
wordcloud~=1.9.4
pillow~=11.1.0
Flask~=3.1.0
SQLAlchemy~=2.0.40
requests~=2.32.0